refactor parts of the library and additional faker implementation
